refactor: remove underscore in jest test files
This will break the build but we will address the breaking changes in
follow-up.
